rename event_hub_client.js
